refactor(dsg-utils): moving the `removeTSClassDeclares` function
