+++ libs/util/code-gen-utils/src/lib/ast/index.ts
@@ -1,2 +1,3 @@
 export * from "./remove-eslint-comments";
+export * from "./remove-typescript-class-declares";
 export * from "./remove-typescript-variable-declares";
