+++ libs/util/code-gen-utils/src/lib/ast/remove-typescript-class-declares/index.ts
@@ -1,0 +1,1 @@
+export * from "./main";
