+++ packages/data-service-generator/src/server/resource/controller/create-controller.ts
@@ -3,6 +3,7 @@ import {
   readFile,
   removeESLintComments,
   removeTSVariableDeclares,
+  removeTSClassDeclares,
 } from "@amplication/code-gen-utils";
 import { builders, namedTypes } from "ast-types";
 import { camelCase } from "camel-case";
@@ -26,7 +27,6 @@ import {
   addAutoGenerationComment,
   addImports,
   removeTSInterfaceDeclares,
-  removeTSClassDeclares,
   getClassDeclarationById,
   importContainedIdentifiers,
   getMethods,
