+++ packages/data-service-generator/src/server/resource/test/create-controller-spec.ts
@@ -1,6 +1,10 @@
 import { namedTypes, builders } from "ast-types";
 import * as kinds from "ast-types/gen/kinds";
-import { print, removeTSVariableDeclares } from "@amplication/code-gen-utils";
+import {
+  print,
+  removeTSVariableDeclares,
+  removeTSClassDeclares,
+} from "@amplication/code-gen-utils";
 import { camelCase } from "camel-case";
 import replaceExt from "replace-ext";
 import { ScalarType, ObjectField, ScalarField } from "prisma-schema-dsl-types";
@@ -10,7 +14,6 @@ import {
   interpolate,
   importNames,
   addImports,
-  removeTSClassDeclares,
   removeTSInterfaceDeclares,
   removeTSIgnoreComments,
 } from "../../../utils/ast";
